feat: Eliminacion producto
Proteccion al eliminar un producto, de no eliminar la lista de opciones si la esta ocupando algun otro

+++ src/schemas/productOptionGroup.js
@@ -1,5 +1,6 @@
 import { model, Schema } from 'mongoose';
 import ProductOption from '#src/schemas/productOption';
+import Product from '#src/schemas/product';
 
 const ProductOptionGroupSchema = new Schema(
   {
@@ -24,6 +25,12 @@ ProductOptionGroupSchema.pre(
   'deleteOne',
   { document: true, query: false },
   async function (next) {
+    const existingInProducts = await Product.find({
+      'options': { $in: [this._id] }
+    })
+    if (existingInProducts.length > 0) {
+      throw new Error('No se puede eliminar el grupo de opciones porque está siendo usado por un producto');
+    }
     for (const option of this.options) {
       const deletedProductOption = ProductOption.findById(option);
       await deletedProductOption.deleteOne();
